+++ packages/eslint-plugin/src/rules/no-base-to-string.ts
@@ -112,6 +112,18 @@ export default util.createRule<Options, MessageIds>({
         return Usefulness.Always;
       }
 
+      if (type.isIntersection()) {
+        for (const subType of type.types) {
+          const subtypeUsefulness = collectToStringCertainty(subType);
+
+          if (subtypeUsefulness === Usefulness.Always) {
+            return Usefulness.Always;
+          }
+        }
+
+        return Usefulness.Never;
+      }
+
       if (!type.isUnion()) {
         return Usefulness.Never;
       }