+++ packages/types/src/ast-node-types.ts
@@ -1,4 +1,4 @@
-export enum AST_NODE_TYPES {
+enum AST_NODE_TYPES {
   ArrayExpression = 'ArrayExpression',
   ArrayPattern = 'ArrayPattern',
   ArrowFunctionExpression = 'ArrowFunctionExpression',
@@ -163,20 +163,4 @@ export enum AST_NODE_TYPES {
   TSVoidKeyword = 'TSVoidKeyword',
 }
 
-export enum AST_TOKEN_TYPES {
-  Boolean = 'Boolean',
-  Identifier = 'Identifier',
-  JSXIdentifier = 'JSXIdentifier',
-  JSXText = 'JSXText',
-  Keyword = 'Keyword',
-  Null = 'Null',
-  Numeric = 'Numeric',
-  Punctuator = 'Punctuator',
-  RegularExpression = 'RegularExpression',
-  String = 'String',
-  Template = 'Template',
-
-  // comment types
-  Block = 'Block',
-  Line = 'Line',
-}
+export { AST_NODE_TYPES };