+++ packages/types/src/ast-token-types.ts
@@ -1,0 +1,19 @@
+enum AST_TOKEN_TYPES {
+  Boolean = 'Boolean',
+  Identifier = 'Identifier',
+  JSXIdentifier = 'JSXIdentifier',
+  JSXText = 'JSXText',
+  Keyword = 'Keyword',
+  Null = 'Null',
+  Numeric = 'Numeric',
+  Punctuator = 'Punctuator',
+  RegularExpression = 'RegularExpression',
+  String = 'String',
+  Template = 'Template',
+
+  // comment types
+  Block = 'Block',
+  Line = 'Line',
+}
+
+export { AST_TOKEN_TYPES };